Move reusable blocks logic to separate package
