+++ packages/editor/src/index.js
@@ -6,6 +6,7 @@ import '@wordpress/blocks';
 import '@wordpress/core-data';
 import '@wordpress/keyboard-shortcuts';
 import '@wordpress/notices';
+import '@wordpress/reusable-blocks';
 import '@wordpress/rich-text';
 import '@wordpress/viewport';
