+++ packages/editor/src/store/reducer.native.js
@@ -18,7 +18,6 @@ import {
 	saving,
 	postLock,
 	postSavingLock,
-	reusableBlocks,
 	template,
 	isReady,
 	editorSettings,
@@ -95,7 +94,6 @@ export default optimist(
 		saving,
 		postLock,
 		postSavingLock,
-		reusableBlocks,
 		template,
 		isReady,
 		editorSettings,
