+++ packages/reusable-blocks/CHANGELOG.md
@@ -1,0 +1,7 @@
+<!-- Learn how to maintain this file at https://github.com/WordPress/gutenberg/tree/master/packages#maintaining-changelogs. -->
+
+## Unreleased
+
+### Internal
+
+- Reusable block utilities moved from `@wordpress/editor` to `@wordpress/reusable-blocks`.
