+++ packages/reusable-blocks/src/index.js
@@ -1,0 +1,13 @@
+/**
+ * WordPress dependencies
+ */
+import '@wordpress/block-editor';
+import '@wordpress/core-data';
+import '@wordpress/notices';
+
+/**
+ * Internal dependencies
+ */
+import './store';
+
+export { storeConfig } from './store';
