+++ packages/reusable-blocks/src/index.native.js
@@ -1,0 +1,9 @@
+/**
+ * WordPress dependencies
+ */
+import '@wordpress/core-data';
+
+/**
+ * Internal dependencies
+ */
+import './store';
