+++ packages/reusable-blocks/src/store/actions.js
@@ -1,0 +1,113 @@
+/**
+ * External dependencies
+ */
+import { castArray } from 'lodash';
+
+/**
+ * Returns an action object used to fetch a single reusable block or all
+ * reusable blocks from the REST API into the store.
+ *
+ * @param {?string} id If given, only a single reusable block with this ID will
+ *                     be fetched.
+ *
+ * @return {Object} Action object.
+ */
+export function __experimentalFetchReusableBlocks( id ) {
+	return {
+		type: 'FETCH_REUSABLE_BLOCKS',
+		id,
+	};
+}
+
+/**
+ * Returns an action object used in signalling that reusable blocks have been
+ * received. `results` is an array of objects containing:
+ *  - `reusableBlock` - Details about how the reusable block is persisted.
+ *  - `parsedBlock` - The original block.
+ *
+ * @param {Object[]} results Reusable blocks received.
+ *
+ * @return {Object} Action object.
+ */
+export function __experimentalReceiveReusableBlocks( results ) {
+	return {
+		type: 'RECEIVE_REUSABLE_BLOCKS',
+		results,
+	};
+}
+
+/**
+ * Returns an action object used to save a reusable block that's in the store to
+ * the REST API.
+ *
+ * @param {Object} id The ID of the reusable block to save.
+ *
+ * @return {Object} Action object.
+ */
+export function __experimentalSaveReusableBlock( id ) {
+	return {
+		type: 'SAVE_REUSABLE_BLOCK',
+		id,
+	};
+}
+
+/**
+ * Returns an action object used to delete a reusable block via the REST API.
+ *
+ * @param {number} id The ID of the reusable block to delete.
+ *
+ * @return {Object} Action object.
+ */
+export function __experimentalDeleteReusableBlock( id ) {
+	return {
+		type: 'DELETE_REUSABLE_BLOCK',
+		id,
+	};
+}
+
+/**
+ * Returns an action object used in signalling that a reusable block is
+ * to be updated.
+ *
+ * @param {number} id      The ID of the reusable block to update.
+ * @param {Object} changes The changes to apply.
+ *
+ * @return {Object} Action object.
+ */
+export function __experimentalUpdateReusableBlock( id, changes ) {
+	return {
+		type: 'UPDATE_REUSABLE_BLOCK',
+		id,
+		changes,
+	};
+}
+
+/**
+ * Returns an action object used to convert a reusable block into a static
+ * block.
+ *
+ * @param {string} clientId The client ID of the block to attach.
+ *
+ * @return {Object} Action object.
+ */
+export function __experimentalConvertBlockToStatic( clientId ) {
+	return {
+		type: 'CONVERT_BLOCK_TO_STATIC',
+		clientId,
+	};
+}
+
+/**
+ * Returns an action object used to convert a static block into a reusable
+ * block.
+ *
+ * @param {string} clientIds The client IDs of the block to detach.
+ *
+ * @return {Object} Action object.
+ */
+export function __experimentalConvertBlockToReusable( clientIds ) {
+	return {
+		type: 'CONVERT_BLOCK_TO_REUSABLE',
+		clientIds: castArray( clientIds ),
+	};
+}
