+++ packages/reusable-blocks/src/store/constants.js
@@ -1,0 +1,14 @@
+/**
+ * Set of post properties for which edits should assume a merging behavior,
+ * assuming an object value.
+ *
+ * @type {Set}
+ */
+export const EDIT_MERGE_PROPERTIES = new Set( [ 'meta' ] );
+
+/**
+ * Constant for the store module (or reducer) key.
+ *
+ * @type {string}
+ */
+export const STORE_KEY = 'core/reusable-blocks';
