+++ packages/reusable-blocks/src/store/index.js
@@ -2,7 +2,6 @@
  * WordPress dependencies
  */
 import { registerStore } from '@wordpress/data';
-import { controls as dataControls } from '@wordpress/data-controls';
 
 /**
  * Internal dependencies
@@ -10,11 +9,11 @@ import { controls as dataControls } from '@wordpress/data-controls';
 import reducer from './reducer';
 import * as selectors from './selectors';
 import * as actions from './actions';
-import controls from './controls';
 import { STORE_KEY } from './constants';
+import applyMiddlewares from './middlewares';
 
 /**
- * Post editor data store configuration.
+ * Data store configuration.
  *
  * @see https://github.com/WordPress/gutenberg/blob/master/packages/data/README.md#registerStore
  *
@@ -24,15 +23,12 @@ export const storeConfig = {
 	reducer,
 	selectors,
 	actions,
-	controls: {
-		...dataControls,
-		...controls,
-	},
 };
 
 const store = registerStore( STORE_KEY, {
 	...storeConfig,
 	persist: [ 'preferences' ],
 } );
+applyMiddlewares( store );
 
 export default store;
