+++ packages/reusable-blocks/src/store/reducer.js
@@ -1,0 +1,103 @@
+/**
+ * External dependencies
+ */
+import optimist from 'redux-optimist';
+import { omit, keyBy } from 'lodash';
+
+/**
+ * WordPress dependencies
+ */
+import { combineReducers } from '@wordpress/data';
+
+export const reusableBlocks = combineReducers( {
+	data( state = {}, action ) {
+		switch ( action.type ) {
+			case 'RECEIVE_REUSABLE_BLOCKS': {
+				return {
+					...state,
+					...keyBy( action.results, 'id' ),
+				};
+			}
+
+			case 'UPDATE_REUSABLE_BLOCK': {
+				const { id, changes } = action;
+				return {
+					...state,
+					[ id ]: {
+						...state[ id ],
+						...changes,
+					},
+				};
+			}
+
+			case 'SAVE_REUSABLE_BLOCK_SUCCESS': {
+				const { id, updatedId } = action;
+
+				// If a temporary reusable block is saved, we swap the temporary id with the final one
+				if ( id === updatedId ) {
+					return state;
+				}
+
+				const value = state[ id ];
+				return {
+					...omit( state, id ),
+					[ updatedId ]: {
+						...value,
+						id: updatedId,
+					},
+				};
+			}
+
+			case 'REMOVE_REUSABLE_BLOCK': {
+				const { id } = action;
+				return omit( state, id );
+			}
+		}
+
+		return state;
+	},
+
+	isFetching( state = {}, action ) {
+		switch ( action.type ) {
+			case 'FETCH_REUSABLE_BLOCKS': {
+				const { id } = action;
+				if ( ! id ) {
+					return state;
+				}
+
+				return {
+					...state,
+					[ id ]: true,
+				};
+			}
+
+			case 'FETCH_REUSABLE_BLOCKS_SUCCESS':
+			case 'FETCH_REUSABLE_BLOCKS_FAILURE': {
+				const { id } = action;
+				return omit( state, id );
+			}
+		}
+
+		return state;
+	},
+
+	isSaving( state = {}, action ) {
+		switch ( action.type ) {
+			case 'SAVE_REUSABLE_BLOCK':
+				return {
+					...state,
+					[ action.id ]: true,
+				};
+
+			case 'SAVE_REUSABLE_BLOCK_SUCCESS':
+			case 'SAVE_REUSABLE_BLOCK_FAILURE': {
+				const { id } = action;
+				return omit( state, id );
+			}
+		}
+
+		return state;
+	},
+} );
+
+export default optimist( reusableBlocks );
